Refactor files and fix errors
